added progress bar, fixed clipToPadding, fixed click
