--- app/src/main/java/com/mayarafernandes/movieplayer/movieList/Movie.kt
@@ -1,117 +1,0 @@
-package com.mayarafernandes.movieplayer.movieList
-
-data class Movie(val title: String,
-                 val description: String,
-                 val type: String,
-                 val publishedDate: Double,
-                 val availableDate: Double,
-                 val metadata: List<MovieMetadata>,
-                 val contents: List<Content>,
-                 val credits: List<Credit>,
-                 val parentalRatings: List<ParentalRating>,
-                 val images: List<Image>,
-                 val categories: List<Category>,
-                 val id: String)
-
-/* object Supplier {
-
-    val movies = listOf<Movie>(
-        Movie(
-            "Amnesia",
-            "Filme confuso do Nolan",
-            "https://upload.wikimedia.org/wikipedia/pt/thumb/c/cf/MementoCartaz.jpg/250px-MementoCartaz.jpg",
-            1
-        ),
-        Movie(
-            "Moana",
-            "Animacao da Disney de 2017",
-            "https://assets.xtechcommerce.com/uploads/images/medium/cdc3456c1b88331938c4102d9221ab2c.jpg",
-            2
-        ),
-        Movie(
-            "IT",
-            "Filme de terror de 2017 baseado no livro homonimo de Stephen King. Conta a historia de crianças de uma cidade lidando com um palhaco que retorna a cada X anos para se alimentar de criancas",
-            "https://www.indiewire.com/wp-content/uploads/2017/08/it.jpg?w=640",
-            3
-        ),
-        Movie(
-            "Rede Social",
-            "Conta a historia de criacao do Facebook.",
-            "https://cdn1.thr.com/sites/default/files/imagecache/portrait_300x450/2010/10/the_social_network_poster_a.jpg",
-            4
-        ),
-        Movie(
-            "Amnesia",
-            "Filme confuso do Nolan",
-            "https://upload.wikimedia.org/wikipedia/pt/thumb/c/cf/MementoCartaz.jpg/250px-MementoCartaz.jpg",
-            5
-        ),
-        Movie(
-            "Moana",
-            "Animacao da Disney de 2017",
-            "https://assets.xtechcommerce.com/uploads/images/medium/cdc3456c1b88331938c4102d9221ab2c.jpg",
-            6
-        ),
-        Movie(
-            "IT",
-            "Filme de terror de 2017 baseado no livro homonimo de Stephen King. Conta a historia de crianças de uma cidade lidando com um palhaco que retorna a cada X anos para se alimentar de criancas",
-            "https://www.indiewire.com/wp-content/uploads/2017/08/it.jpg?w=640",
-            7
-        ),
-        Movie(
-            "Rede Social",
-            "Conta a historia de criacao do Facebook.",
-            "https://cdn1.thr.com/sites/default/files/imagecache/portrait_300x450/2010/10/the_social_network_poster_a.jpg",
-            8
-        ),
-        Movie(
-            "Amnesia",
-            "Filme confuso do Nolan",
-            "https://upload.wikimedia.org/wikipedia/pt/thumb/c/cf/MementoCartaz.jpg/250px-MementoCartaz.jpg",
-            9
-        ),
-        Movie(
-            "Moana",
-            "Animacao da Disney de 2017",
-            "https://assets.xtechcommerce.com/uploads/images/medium/cdc3456c1b88331938c4102d9221ab2c.jpg",
-            10
-        ),
-        Movie(
-            "IT",
-            "Filme de terror de 2017 baseado no livro homonimo de Stephen King. Conta a historia de crianças de uma cidade lidando com um palhaco que retorna a cada X anos para se alimentar de criancas",
-            "https://www.indiewire.com/wp-content/uploads/2017/08/it.jpg?w=640",
-            11
-        ),
-        Movie(
-            "Rede Social",
-            "Conta a historia de criacao do Facebook.",
-            "https://cdn1.thr.com/sites/default/files/imagecache/portrait_300x450/2010/10/the_social_network_poster_a.jpg",
-            12
-        ),
-        Movie(
-            "Amnesia",
-            "Filme confuso do Nolan",
-            "https://upload.wikimedia.org/wikipedia/pt/thumb/c/cf/MementoCartaz.jpg/250px-MementoCartaz.jpg",
-            13
-        ),
-        Movie(
-            "Moana",
-            "Animacao da Disney de 2017",
-            "https://assets.xtechcommerce.com/uploads/images/medium/cdc3456c1b88331938c4102d9221ab2c.jpg",
-            14
-        ),
-        Movie(
-            "IT",
-            "Filme de terror de 2017 baseado no livro homonimo de Stephen King. Conta a historia de crianças de uma cidade lidando com um palhaco que retorna a cada X anos para se alimentar de criancas",
-            "https://www.indiewire.com/wp-content/uploads/2017/08/it.jpg?w=640",
-            15
-        ),
-        Movie(
-            "Rede Social",
-            "Conta a historia de criacao do Facebook.",
-            "https://cdn1.thr.com/sites/default/files/imagecache/portrait_300x450/2010/10/the_social_network_poster_a.jpg",
-            16
-        )
-    )
-}
-*/
