+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/MovieRepository.kt
@@ -15,7 +15,6 @@ class MovieRepository (private val localMovieStorage: LocalMovieStorage) {
             }
 
             override fun onError() { movieCallbacks.onError() }
-
         })
     }
