+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/service/MovieCallbacks.kt
@@ -3,7 +3,6 @@ package com.mayarafernandes.movieplayer.movieList.service
 import com.mayarafernandes.movieplayer.movieList.Movie
 
 interface MovieCallbacks {
-
     fun onSuccess(movieList: List<Movie>)
     fun onError()
 }
