+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/service/MovieServiceIMPL.kt
@@ -57,5 +57,4 @@ class MovieServiceIMPL() {
         movie.images,
         movie.categories,
         movie.id)
-
 }
