+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/view/MovieListActivity.kt
@@ -2,6 +2,8 @@ package com.mayarafernandes.movieplayer.movieList.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.View
+import android.widget.ProgressBar
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.mayarafernandes.movieplayer.R
 import com.mayarafernandes.movieplayer.movieList.*
@@ -13,10 +15,13 @@ class MovieListActivity : AppCompatActivity(),
     MovieListView, MovieViewModelClickListener {
 
     private lateinit var controller: MovieListController
+    private lateinit var progressBar: ProgressBar
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_movie_list)
+
+        progressBar = findViewById(R.id.movieListLoadingProgressBar)
 
         val memoryRepository = MemoryRepository()
         val movieRepository =
@@ -28,7 +33,7 @@ class MovieListActivity : AppCompatActivity(),
         layoutManager.orientation = LinearLayoutManager.VERTICAL
         movieListRecyclerView.layoutManager = layoutManager
 
-        val rvAdapter = MovieListRecyclerViewAdapter(this, this)
+        val rvAdapter = MovieListRecyclerViewAdapter(this, this, progressBar)
         movieListRecyclerView.adapter = rvAdapter
 
         controller.onViewCreated()
@@ -41,5 +46,10 @@ class MovieListActivity : AppCompatActivity(),
 
     override fun onClick(movie: MovieViewModel) {
         controller.onSelectMovie(movie)
+    }
+
+    override fun showProgressBar(show: Boolean) {
+        if(show) progressBar.visibility = View.VISIBLE
+        else progressBar.visibility = View.GONE
     }
 }
