+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/view/MovieListView.kt
@@ -1,8 +1,6 @@
 package com.mayarafernandes.movieplayer.movieList.view
 
-import com.mayarafernandes.movieplayer.movieList.view.MovieViewModel
-
 interface MovieListView {
-
     fun setViewModel(viewModels: List<MovieViewModel>)
+    fun showProgressBar(show: Boolean)
 }
