+++ app/src/main/res/layout/adapter_movie_item.xml
@@ -12,8 +12,8 @@
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="397dp"
-            android:layout_height="wrap_content">
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
 
         <ImageView
                 android:id="@+id/movieIconImageView"
@@ -30,12 +30,16 @@
 
         <TextView
                 android:id="@+id/movieTitleTextView"
-                android:layout_width="wrap_content"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="16dp"
                 android:layout_marginTop="16dp"
+                android:layout_marginEnd="16dp"
+                android:autoSizeTextType="uniform"
                 android:text="@string/movieTitle"
-                android:textSize="24sp"
+                android:textStyle="bold"
+                android:maxLines="2"
+                android:ellipsize="end"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.018"
                 app:layout_constraintStart_toEndOf="@+id/movieIconImageView"
@@ -46,10 +50,12 @@
                 android:layout_width="0dp"
                 android:layout_height="0dp"
                 android:layout_marginTop="16dp"
-                android:layout_marginEnd="16dp"
                 android:text="@string/movieDescription"
+                android:autoSizeTextType="uniform"
+                android:maxLines="3"
+                android:ellipsize="end"
                 app:layout_constraintBottom_toBottomOf="@+id/movieIconImageView"
-                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintEnd_toEndOf="@+id/movieTitleTextView"
                 app:layout_constraintHorizontal_bias="0.0"
                 app:layout_constraintStart_toStartOf="@+id/movieTitleTextView"
                 app:layout_constraintTop_toBottomOf="@+id/movieTitleTextView"
